moved SessionPool options into options.h
